+++ example/lib/interface/knight_interface.dart
@@ -4,6 +4,7 @@ import 'package:example/player/knight.dart';
 import 'package:flutter/material.dart';
 
 class KnightInterface extends GameInterface {
+  OverlayEntry? overlay;
   @override
   Future<void> onLoad() {
     add(BarLifeComponent());
@@ -29,8 +30,36 @@ class KnightInterface extends GameInterface {
       id: 5,
       position: Vector2(200, 20),
       onTapComponent: (selected) {
+        // if (gameRef.player != null) {
+        //   (gameRef.player as Knight).changeControllerToVisibleEnemy();
+        // }
+        if (overlay != null) {
+          overlay?.remove();
+          overlay = null;
+          return;
+        }
         if (gameRef.player != null) {
-          (gameRef.player as Knight).changeControllerToVisibleEnemy();
+          overlay = FollowerWidget.show(
+            context,
+            gameRef.player!,
+            Container(
+              decoration: BoxDecoration(
+                color: Colors.white,
+                borderRadius: BorderRadius.circular(5),
+              ),
+              padding: EdgeInsets.all(10),
+              child: ElevatedButton(
+                onPressed: () {
+                  print('aqui');
+                },
+                child: Text('clica aqui'),
+              ),
+            ),
+            align: Offset(
+              0,
+              -40,
+            ),
+          );
         }
       },
     ));